refactor: 패키지 변경
- View 와 그 구현체가 각각 view 와 view.impl 패키지 하위에 있도록
