+++ mvc/src/main/java/com/interface21/webmvc/servlet/view/impl/JspView.java
@@ -1,6 +1,6 @@
-package com.interface21.webmvc.servlet.view;
+package com.interface21.webmvc.servlet.view.impl;
 
-import com.interface21.webmvc.servlet.View;
+import com.interface21.webmvc.servlet.view.View;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.util.Map;
